Back/requirements.txt
fastapi
pydantic
fastapi-sqlalchemy
sqlalchemy-utils
alembic
uvicorn
jinja2
python-multipart
passlib
pyjwt
psycopg2
babel